Further fixing correct separation of files
